Fixed a few issues to make it compiler under Visual C++ 10

+++ src/hpp/dds/core/status/State.hpp
@@ -32,8 +32,8 @@ public:
 
 public:
     SampleRejectedState() : MaskType() { }
-    SampleRejectedState(const SampleRejectedState& src) : MaskType(src.to_ulong()) { }
-    SampleRejectedState(const MaskType& src) : MaskType(src.to_ulong()) { }
+	SampleRejectedState(const SampleRejectedState& src) : MaskType(static_cast<int>(src.to_ulong())) { }
+    SampleRejectedState(const MaskType& src) : MaskType(static_cast<int>(src.to_ulong())) { }
 
 public:
     inline static const SampleRejectedState not_rejected() {
@@ -69,7 +69,7 @@ public:
 public:
     StatusMask() { }
     explicit StatusMask(uint64_t mask) : std::bitset<OMG_DDS_STATUS_COUNT>(mask) { }
-    StatusMask(const StatusMask& other) : MaskType(other.to_ulong()) { }
+    StatusMask(const StatusMask& other) : MaskType(static_cast<int>(other.to_ulong())) { }
     ~StatusMask() { }
 
 public:
